backend/requirements.txt
openai 
langchain-openai 
langchain_community
langchain-huggingface
fastapi
pandas
nltk
tqdm
pyYAML
python-box
ensure==1.0.4
fastapi
uvicorn
gdown
docx2txt
langchain_experimental
pypdf
chardet
sentence-transformers
huggingface_hub[hf_xet]
python-docx
unstructured
unstructured[local-inference] 
pytesseract
supabase
langchain_google_genai
torch
langgraph

-e .